core/pyproject.toml
[tool.poetry]
name = "dnd_scribe_core"
version = "0.1.0"
description = ""
authors = ["Daomephsta <dtnilfisk@gmail.com>"]
readme = "README.md"
packages = [{include = "dnd_scribe/core"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
markdown = "^3.4.3"
python-frontmatter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
